fix(sudo): improve /sudo command execution logic
The sudo command's logic has been improved: executing commands as
someone is now working (Player#chat) saying a message as somebody else
works by first invoking the AsyncChatEvent, and then taking each
element of the audience and send every individual the message. The
wheel permission was added to paper-plugin.yml to restrict the /sudo
command to admins or staff.

Summary of changes:

- Add `powertools.wheel` permission to `paper-plugin.yml`
- Refactor `executeSudoCommand` to handle command command execution
  better
- Refactor `executeSudoChat` to use `AsyncChatEvent` for better chat
  message handling and send the message to every player